refactor(react-grid): rename VirtualTableView to VirtualTable
