+++ packages/dx-react-demos/src/bootstrap3/featured-virtual-scrolling/demo.jsx
@@ -5,7 +5,7 @@ import {
 } from '@devexpress/dx-react-grid';
 import {
   Grid,
-  VirtualTableView, TableHeaderRow, TableFilterRow, TableSelection, TableGroupRow,
+  VirtualTable, TableHeaderRow, TableFilterRow, TableSelection, TableGroupRow,
   TableColumnReordering, GroupingPanel, DragDropContext,
 } from '@devexpress/dx-react-grid-bootstrap3';
 import {
@@ -85,7 +85,7 @@ export default class Demo extends React.PureComponent {
 
         <SelectionState />
 
-        <VirtualTableView
+        <VirtualTable
           tableCellTemplate={this.tableCellTemplate}
         />
