+++ packages/dx-react-demos/src/bootstrap3/sorting/remote-sorting.jsx
@@ -4,7 +4,7 @@ import {
 } from '@devexpress/dx-react-grid';
 import {
   Grid,
-  VirtualTableView,
+  VirtualTable,
   TableHeaderRow,
 } from '@devexpress/dx-react-grid-bootstrap3';
 
@@ -85,7 +85,7 @@ export default class Demo extends React.PureComponent {
             sorting={sorting}
             onSortingChange={this.changeSorting}
           />
-          <VirtualTableView />
+          <VirtualTable />
           <TableHeaderRow allowSorting />
         </Grid>
         {loading && <Loading />}
