+++ packages/dx-react-demos/src/material-ui/featured-virtual-scrolling/demo.jsx
@@ -5,7 +5,7 @@ import {
 } from '@devexpress/dx-react-grid';
 import {
   Grid,
-  VirtualTableView, TableHeaderRow, TableFilterRow, TableSelection, TableGroupRow,
+  VirtualTable, TableHeaderRow, TableFilterRow, TableSelection, TableGroupRow,
   GroupingPanel, DragDropContext, TableColumnReordering,
 } from '@devexpress/dx-react-grid-material-ui';
 import {
@@ -85,7 +85,7 @@ export default class Demo extends React.PureComponent {
 
         <SelectionState />
 
-        <VirtualTableView
+        <VirtualTable
           tableCellTemplate={this.tableCellTemplate}
         />
         <TableHeaderRow allowSorting allowDragging />
