+++ packages/dx-react-demos/src/material-ui/virtual-scrolling/basic.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {
   Grid,
-  VirtualTableView,
+  VirtualTable,
   TableHeaderRow,
 } from '@devexpress/dx-react-grid-material-ui';
 
@@ -39,7 +39,7 @@ export default class Demo extends React.PureComponent {
         columns={columns}
         getRowId={getRowId}
       >
-        <VirtualTableView />
+        <VirtualTable />
         <TableHeaderRow />
       </Grid>
     );
