+++ packages/dx-react-grid/demos/featured/virtual-scrolling.md
@@ -1,5 +1,5 @@
 # React Grid Virtual Scrolling (200K rows)
 
-The `VirtualTableView` plugin provides the virtual scrolling mode as an alternative to data paging. It allows end users to navigate to data rows using the vertical scrollbar. In this demo, the Grid with enabled virtual scrolling is bound to a 200,000 record data source.
+The `VirtualTable` plugin provides the virtual scrolling mode as an alternative to data paging. It allows end users to navigate to data rows using the vertical scrollbar. In this demo, the Grid with enabled virtual scrolling is bound to a 200,000 record data source.
 
 .embedded-demo(featured-virtual-scrolling/demo)
